refactor: RN-Web 통신방식 1차 구현
