+++ apps/enfpy-native/App.tsx
@@ -1,20 +1,18 @@
 import React from 'react';
 import {StatusBar} from 'react-native';
-import {WebBridgeProvider} from './src/contexts/web-bridge/WebBridgeContext';
 import {MainStackNavigation} from './src/navigation/main-stack/MainStackNavigation';
+import {BridgeProvider} from './src/contexts/web-bridge/BridgeContext';
 
 const App = () => {
   return (
-    <>
+    <BridgeProvider>
       <StatusBar
         backgroundColor="transparent"
         translucent
         barStyle="light-content"
       />
-      <WebBridgeProvider>
-        <MainStackNavigation />
-      </WebBridgeProvider>
-    </>
+      <MainStackNavigation />
+    </BridgeProvider>
   );
 };
