+++ packages/modules/web-bridge/hooks/useBridgeMessageCreator.ts
@@ -1,0 +1,9 @@
+import { useDev } from "./useDev";
+import { useNavigation } from "./useNavigation";
+
+export const useBridgeMessageCreator = () => {
+  const devMessages = useDev();
+  const navigationMessages = useNavigation();
+
+  return { dev: { ...devMessages }, navigation: { ...navigationMessages } };
+};
