+++ packages/modules/web-bridge/hooks/useNavigation.ts
@@ -1,17 +1,19 @@
 import { WebBridgeActionDatas, WebBridgeActions } from "../types/action";
-import { useBridge } from "./useBridge";
+import { postMessage } from "../utils/message";
 
 export const useNavigation = () => {
-  const bridge = useBridge();
-
   const navigate = (
     args: WebBridgeActionDatas[WebBridgeActions.NAVIGATION_NAVIGATE]
   ) => {
-    bridge.postMessage(WebBridgeActions.NAVIGATION_NAVIGATE, args);
+    postMessage(WebBridgeActions.NAVIGATION_NAVIGATE, args);
   };
 
   const goBack = () => {
-    bridge.postMessage(WebBridgeActions.NAVIGATION_GO_BACK);
+    postMessage(WebBridgeActions.NAVIGATION_GO_BACK);
   };
-  return { navigate, goBack };
+
+  return {
+    navigate,
+    goBack,
+  };
 };
